57. new Migration: All PKs have type int

+++ Schedule_App.Core/DTOs/Lesson/LessonCreateDTO.cs
@@ -19,12 +19,12 @@ namespace Schedule_App.Core.DTOs.Lesson
     public class LessonCreateDTO
     {
         [Required]
-        [Range(1, short.MaxValue, ErrorMessage = "The Classroom's id must be a positive integer greater than zero.")]
-        public short ClassroomId { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "The Classroom's id must be a positive integer greater than zero.")]
+        public int ClassroomId { get; set; }
 
         [Required]
-        [Range(1, short.MaxValue, ErrorMessage = "The Subject's id must be a positive integer greater than zero.")]
-        public short SubjectId { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "The Subject's id must be a positive integer greater than zero.")]
+        public int SubjectId { get; set; }
 
         [Required]
         [Range(1, int.MaxValue, ErrorMessage = "The Group's id must be a positive integer greater than zero.")]
@@ -46,7 +46,7 @@ namespace Schedule_App.Core.DTOs.Lesson
         public DateTime EndsAt { get; set; }
 
         [Required]
-        [Range(1, short.MaxValue, ErrorMessage = "The Status's id must be a positive integer greater than zero.")]
-        public short StatusId { get; set; }
+        [Range(1, int.MaxValue, ErrorMessage = "The Status's id must be a positive integer greater than zero.")]
+        public int StatusId { get; set; }
     }
 }
